Refactor reading rate limit annotations
